github(login): Add a way to read ~/.netrc.gpg encrypted files
I usually don't store authentication information in plain text in my
home directory. Instead most of my secrets are stored in *.gpg files
which can be read by relevant applications if my GPG key has been
unlocked in my X session.

The current implementation of travis.rb only reads ~/.netrc for
`--auto` logins. It would be nice for the cli to also be able to read
`~/.netrc.gpg` file.

This PR adds the gpgme gem as a dependency to be able to read *.gpg
files if needs be.

I am not sure requiring the `gpgme` gem in the gemspec is a good idea
as it really depends on wether you have *.gpg files to read or not.
Do you have a better idea on how we could do that?

Thanks!